Created hargrave-writing-sample
This is my 'Writing Exercise' topic for the 'Skill Assessment'. Please note that the 'Bus' and 'Personal vehicle' sections are placeholders, since the assignment was to describe a single form of transportation to Salesforce Tower. Please contact Valkyrie Hunter if whoever first sees this PR has no idea what to do with it! :) Thank you!